chore(repo): undo gradle tests
